field 'yearsOfExperience' Integer to Double

+++ src/main/java/com/fran/jobsy/app/dto/offering/OfferingRequest.java
@@ -21,6 +21,6 @@ public record OfferingRequest(
         @NotNull(message = "Years of experience is mandatory")
         @Min(value = 0, message = "Years of experience cannot be negative")
         @Max(value = 50, message = "Years of experience must not exceed 50")
-        Integer yearsOfExperience
+        Double yearsOfExperience
 ) {
 }
